requirements.txt
openai
telebot
python-dotenv
tiktoken